updated shell script to improve setup

+++ setup.sh
@@ -1,8 +1,31 @@
-python3 -m venv venv
-source venv/bin/activate
-python -m pip install --upgrade pip
-pip install -r requirements-dump.txt
+# Create and activate virtual environment
+if [ ! -d "venv" ]; then
+    python3 -m venv venv
+    echo "Virtual environment 'venv' created."
+else
+    echo "Virtual environment 'venv' already exists."
+fi
 
+# Activate virtual environment and upgrade pip
+source venv/bin/activate
+if [ $? -eq 0 ]; then
+    echo "Virtual environment 'venv' activated."
+    python -m pip install --upgrade pip
+    echo "pip upgraded to the latest version."
+else
+    echo "Error: Failed to activate virtual environment 'venv'."
+    exit 1
+fi
+
+# Check if requirements-dump.txt exists before attempting to install
+if [ -f "requirements-dump.txt" ]; then
+    pip install -r requirements-dump.txt
+else
+    echo "Error: requirements-dump.txt not found."
+    exit 1
+fi
+
+# Check and create DB directory if it doesn't exist
 if [ -d "DB" ]
 then
     echo "Directory 'DB' exists."
@@ -12,6 +35,7 @@ else
     echo "Directory 'DB' created."
 fi
 
+# Check and create .env file if it doesn't exist
 if [ -f ".env" ]
 then
     echo "File '.env' exists."
